⚡️  Store existing migrated versions to save some queries
The current code run the following query for each file to migrate to
check if it was migrated already or not. In this context the `versions`
values could be memoized or loaded in the object initialization.

```ruby
query = Virsandra::SelectQuery.new("name").from(SCHEMA_TABLE)
keyspace.execute(query.to_s).to_a.map(&:values).flatten
```